Chore: merge all html formatter files into `html.js` (#14612)
